chore: rename golang -> go (#2024)
because no one actually says golang
