+++ packages/jsii-pacmak/lib/targets/go/emit-context.ts
