+++ packages/jsii-pacmak/lib/targets/go/package.ts
@@ -6,7 +6,7 @@ import { EmitContext } from './emit-context';
 import { GoClass, GoType, Enum, Interface, Struct } from './types';
 import { findTypeInTree, goPackageName, flatMap } from './util';
 
-// JSII golang runtime module name
+// JSII go runtime module name
 const JSII_MODULE_NAME = 'github.com/aws-cdk/jsii/jsii-experimental';
 
 /*
