+++ packages/jsii-pacmak/lib/targets/go/runtime.ts
