+++ packages/jsii-pacmak/lib/targets/go/types/class.ts
