+++ packages/jsii-pacmak/lib/targets/go/types/enum.ts
