+++ packages/jsii-pacmak/lib/targets/go/types/go-type-reference.ts
