+++ packages/jsii-pacmak/lib/targets/go/types/index.ts
