+++ packages/jsii-pacmak/lib/targets/go/types/interface.ts
