+++ packages/jsii-pacmak/lib/targets/go/types/struct.ts
