+++ packages/jsii-pacmak/lib/targets/go/types/type-member.ts
