+++ packages/jsii-pacmak/lib/targets/go/util.ts
